refactor!: major project restructuring (CLI, parsing, conditions, logging)
