+++ src/parsing/__init__.py
@@ -1,0 +1,27 @@
+from .lexer import (
+    ConfigSections,
+    normalize_expression,
+    parse_config_sections,
+    tokenize_expression,
+)
+from .parser import (
+    build_conditions,
+    build_context,
+    build_fact_condition,
+    build_query_condition,
+    build_rule_condition,
+    parse_expression,
+)
+
+__all__ = [
+    "ConfigSections",
+    "normalize_expression",
+    "parse_config_sections",
+    "tokenize_expression",
+    "build_conditions",
+    "build_context",
+    "build_fact_condition",
+    "build_query_condition",
+    "build_rule_condition",
+    "parse_expression",
+]
